+++ oneid/authentication.py
@@ -7,6 +7,7 @@ from drf_expiring_authtoken.authentication import ExpiringTokenAuthentication
 from django.conf import settings
 
 from oneid_meta.models import User
+from oneid.statistics import UserStatistics
 
 
 class HeaderArkerBaseAuthentication(BaseAuthentication):
@@ -34,23 +35,7 @@ class HeaderArkerBaseAuthentication(BaseAuthentication):
 
 class CustomExpiringTokenAuthentication(ExpiringTokenAuthentication):
     def authenticate_credentials(self, key):
-        import datetime
-        import redis
-        user, token = super().authenticate_credentials(key)
-        date = datetime.datetime.today().date().isoformat()
-        if hasattr(token.user, 'uuid'):
-            uuid = str(token.user.uuid)
-            redis_conn = redis.Redis(settings.REDIS_HOST)
-            key = 'active-' + date
-            res = redis_conn.hgetall(key)
 
-            if res:
-                test_uuid = redis_conn.hget(key, uuid)
-                if not test_uuid:
-                    redis_conn.hincrby(key, uuid, 1)
-                else:
-                    redis_conn.hincrby(key, uuid, 1)
-            else:
-                redis_conn.hset(key, uuid, 1)
-                redis_conn.expire(key, settings.ACTIVE_USER_DATA_LIFEDAY * 60 * 60 * 24)
+        user, token = super().authenticate_credentials(key)
+        UserStatistics.set_active_count(user)
         return user, token
